# Improve user ergonomics, performance and deny a timing attack