--- a/benches/jwt.rs
+++ b/benches/jwt.rs
@@ -23,10 +23,6 @@ fn bench_encode(b: &mut test::Bencher) {
 
 #[bench]
 fn bench_decode(b: &mut test::Bencher) {
-    let token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ".to_owned();
-    b.iter(|| decode::<Claims>(
-        token.clone(),
-        "secret".to_owned(),
-        Algorithm::HS256
-    ));
+    let token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ";
+    b.iter(|| decode::<Claims>(token, "secret", Algorithm::HS256));
 }