(sorry for the bundled commits, I can separate them into separate PRs if you need)

This looks pretty useful so I wanted to help improve the ergonomics a bit (and then got a bit sidetracked in making it faster). Most of the ownership was not required (so `to_owned` is generally unnecessary), as the majority of work cannot be done in-place (at least not with current libraries).

This commit fixes a potential timing attack on hmac verification: 486c4a87f9603250167e40d70da2845e04a07500

Performance change:
```bash
# before
test bench_decode ... bench:      11,501 ns/iter (+/- 885)
test bench_encode ... bench:       5,036 ns/iter (+/- 511)

# after
test bench_decode ... bench:       3,460 ns/iter (+/- 718)
test bench_encode ... bench:       3,964 ns/iter (+/- 516)
```